修改微信支付金额单位(元)

+++ src/Magicodes.Pay.Wxpay/Pay/Dto/AppPayInput.cs
@@ -48,7 +48,7 @@ namespace Magicodes.Pay.Wxpay.Pay.Dto
         /// <summary>
         ///     订单总金额，单位为元
         /// </summary>
-        public decimal TotalFee { get; set; }
+        public int TotalFee { get; set; }
 
         /// <summary>
         ///     用户端实际ip
